requirements.txt
black==20.8b1
flake8==3.8.4
gpiozero==1.5.1
mypy==0.790
pycodestyle==2.6.0
Pykka==2.0.2
pytest==6.1.2
RPi.GPIO==0.7.0
thespian==3.10.1
websocket-client==0.57.0
requests==2.25.0
